Update S3 instructions in README.md
Hi Team,
I made an update suggestion at the S3 configuration instructions. While I was following the instructions for my Mentor assessment I missed the Versioning and it caused some "problems" for me after for rolling back the modifications.
It's just a suggestion so feel free to assume this is appropriated or not.

Thank you very much!